% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{add_future_net_dist}
\alias{add_future_net_dist}
\title{Title Add future net input distribution}
\usage{
add_future_net_dist(
  interventions,
  group_var,
  off_year_max = 0.2,
  cycle_period = 3
)
}
\arguments{
\item{interventions}{Site file interventions section}

\item{group_var}{Site grouping}

\item{off_year_max}{Maximum distribution outside of N-yearly mass}

\item{cycle_period}{Assumed cycle period (years) for mass distribution}
}
\description{
Infer cycle period (for N-yearly main distribution frequency) and fit input
distribution, ensuring already calibrated values remain fixed). Requires `itn_input_dist`
}
