% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_df}
\alias{aggregate_df}
\title{Aggregate a data frame}
\usage{
aggregate_df(
  df,
  groups,
  sum_cols = NULL,
  mean_cols = NULL,
  weighted_mean_cols = NULL,
  w = NULL
)
}
\arguments{
\item{df}{data.frame}

\item{groups}{character vector of grouping variables}

\item{sum_cols}{character vector of variables to summarise by sum}

\item{mean_cols}{character vector of variables to summarise by mean}

\item{weighted_mean_cols}{character vector of variables to summarise by weighted mean}

\item{w}{column name for weights to be used for weighted mean summary}
}
\value{
An aggregated (un-grouped) data frame
}
\description{
Aggregate a data frame
}
