% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_normExpr.Droplet}
\alias{calc_normExpr.Droplet}
\title{Calculate normalized expresssion for UMI-based data}
\usage{
calc_normExpr.Droplet(x, factor = 10000, logarithmetic = T)
}
\arguments{
\item{x}{Matrix or sparse matrix. Each column specify a sample.}

\item{factor}{A number define the resulting library size for each sample after the normalization. Default 10000.}

\item{logarithmetic}{Logical. Should the results be log-transformed?}
}
\value{
Matrix
}
\description{
'calc_normExpr.Droplet' is a wrapper to calculate the normalized expression for droplet data.
It would calculate the normalized expresssion as: 'x/sum(x) * factor' for each sample.
}
