% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_TPM}
\alias{calc_TPM}
\title{Calculate TPM}
\usage{
calc_TPM(x)
}
\arguments{
\item{x}{Named vector of gene counts}
}
\value{
Named vector of TPM
}
\description{
Calculate TPM, the length information is provided by R package \code{annotables}
}
