% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcCorrelative2}
\alias{calcCorrelative2}
\title{Check consensus program values.ver2}
\usage{
calcCorrelative2(
  program_mt,
  sources,
  celltypes,
  nonzero_ratio = 0.1,
  rho_cutoff = 0,
  pval_cutoff = 0.1
)
}
\arguments{
\item{program_mt}{Matrix, processed}

\item{sources}{Named strings, c\('sample_name' = 'bulk|singlecell'\)}

\item{celltypes}{Named vector, should be a cell-id-named vector of cell types}

\item{nonzero_ratio}{Value between 0-1, c\('sample_name' = 'bulk|singlecell'\)}
}
\value{
A named list with the correlative information between celltypes.
}
\description{
Check consensus program values.ver2
}
