% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcCorrelative3}
\alias{calcCorrelative3}
\title{Check consensus program values.ver3}
\usage{
calcCorrelative3(
  program_mt,
  sources,
  celltypes,
  nonzero_ratio = 0.1,
  rho_cutoff = 0,
  n_permtest = 2000,
  n_cores = parallel::detectCores() - 1,
  qval_cutoff = 0.05
)
}
\arguments{
\item{program_mt}{Matrix, processed}

\item{sources}{Named strings, c\('sample_name' = 'bulk|singlecell'\)}

\item{celltypes}{Named vector, should be a cell-id-named vector of cell types}

\item{nonzero_ratio}{Value between 0-1, c\('sample_name' = 'bulk|singlecell'\)}

\item{rho_cutoff}{The remaining program pairs must have spearman correlation rho larger than this. Default = 0.}

\item{n_permtest}{Number of permutation rounds to perform, default = 2000}

\item{n_cores}{Number of cores to use for permutation, default = the maximum available cores - 1}

\item{qval_cutoff}{The remaining program pairs must have qvalue smaller than this. Default = 0.05. BH method.}
}
\value{
A named list with the correlative information between celltypes.
}
\description{
Check consensus program values.ver3
}
