% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkCorrelative}
\alias{checkCorrelative}
\title{Check and plot the correlation between two programs}
\usage{
checkCorrelative(x_label, y_label, test_obj, plt_obj, ...)
}
\arguments{
\item{x_label}{Name of a program, x}

\item{y_label}{Name of a program, y}

\item{test_obj}{A test object}

\item{...}{Options pass to Seurat::FeaturePlot}
}
\value{
Ggplot objects
}
\description{
Check and plot the correlation between two programs
}
