% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkCorrelative2}
\alias{checkCorrelative2}
\title{Check and plot the correlation between two programs.ver2}
\usage{
checkCorrelative2(x_label, y_label, test_obj, plt_obj, ...)
}
\arguments{
\item{x_label}{Name of a program, x}

\item{y_label}{Name of a program, y}

\item{test_obj}{A test object}

\item{plt_obj}{An object for ploting generated by \code{checkConsensusProgram}}

\item{...}{Other options}
}
\value{
ggplot objects
}
\description{
Check and plot the correlation between two programs.ver2
}
