% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{checkCorrelative_context}
\alias{checkCorrelative_context}
\title{Check and plot the correlation between two programs within specific context}
\usage{
checkCorrelative_context(x_label, y_label, test_obj_context, plt_obj, ...)
}
\arguments{
\item{x_label}{Name of a program, x}

\item{y_label}{Name of a program, y}

\item{test_obj_context}{A test object with multiple contexts}

\item{plt_obj}{An object for ploting generated by \code{checkConsensusProgram}}

\item{...}{Options pass to Seurat::FeaturePlot}
}
\value{
ggplot objects
}
\description{
Check and plot the correlation between two programs within specific context
}
