% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complementCor}
\alias{complementCor}
\title{Generate a random variable with a defined correlation to an existing variables
This function is adapted from the \href{https://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variables}{post}}
\usage{
complementCor(y, rho, x)
}
\arguments{
\item{y}{The reference vector.}

\item{rho}{The target correlation}

\item{x}{A vector to start with. If specified, it must have the same length as \code{y}}
}
\description{
Generate a random variable with a defined correlation to an existing variables
This function is adapted from the \href{https://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variables}{post}
}
