% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dea.wilcox}
\alias{dea.wilcox}
\title{Differential expression analysis using wilcox test}
\usage{
dea.wilcox(
  mt1_origin,
  mt2_origin,
  down_size = 5000,
  n_cores = 1,
  onlyPos = T,
  ...
)
}
\description{
Differential expression analysis using wilcox test
}
