% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determineK}
\alias{determineK}
\title{Determine K to use}
\usage{
determineK(data_obj, draw_now = F, verbose = T, ...)
}
\arguments{
\item{data_obj}{Data object from \code{determineK_runNMF}}

\item{draw_now}{Logical. If print out the plots.}

\item{verbose}{Logical. If print the logs.}
}
\value{
A named list with optimum K and data/objects to support it.
}
\description{
Determine the knee point using \code{kneer} to identify the optimum K to run cNMF.
}
