% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determineK_runNMF}
\alias{determineK_runNMF}
\title{Run NMF for determination of K}
\usage{
determineK_runNMF(
  merge_mt,
  Ks2run,
  n_reps,
  n_cores = 0,
  draw_now = F,
  verbose = T,
  return_res = F
)
}
\arguments{
\item{merge_mt}{Matrix or sparse matrix. Normalized expression matrix of single-cell and bulk. Genes x Samples.}

\item{Ks2run}{Integer vector. Ks to test.}

\item{n_reps}{Integer. Number of replication for each K.}

\item{n_cores}{Integer. Number of cores to use, recommended to use multicore. Default = 0, all cores possible.}

\item{draw_now}{Logical. If print out the plots.}

\item{verbose}{Logical. If print the logs.}

\item{return_res}{Logical. If return all results.}
}
\value{
A named list with data to determine K
}
\description{
A wrapper function to run NMF multiple times and calculate the index matrix helping determination of K
}
