% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{encodeContext}
\alias{encodeContext}
\title{Encode context}
\usage{
encodeContext(annotation, context_num = NULL)
}
\arguments{
\item{annotation}{named vector, indicate the context of each bulk sample}

\item{context_num_vec}{named numeric vector, indicate the number assign to each context, to perform multi-variable linear regression}
}
\value{
a tibble
}
\description{
This function remove NA, and assign numbers to each
context, in order to perform multi-variable linear regression
}
