% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{ever_used}
\alias{ever_used}
\title{Select areas that have ever used an intervention}
\usage{
ever_used(interventions, var, group_var)
}
\arguments{
\item{interventions}{Site file interventions section}

\item{var}{Interventions variable to check if ever used}

\item{group_var}{Site grouping}
}
\description{
Select areas that have ever used an intervention
}
