% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_site}
\alias{example_site}
\title{Example site}
\format{
## `example_site`
A named list of length 3.
}
\usage{
example_site
}
\description{
An example site for demonstration
}
\keyword{datasets}
