% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{expand_interventions}
\alias{expand_interventions}
\title{Expand the inereventions time frame for all sites}
\usage{
expand_interventions(interventions, max_year, group_var)
}
\arguments{
\item{interventions}{Site file interventions section}

\item{max_year}{Maximum year to expand to}

\item{group_var}{Site grouping}
}
\description{
Expand the inereventions time frame for all sites
}
