% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractProgramData}
\alias{extractProgramData}
\title{Extract and scale the data for ploting}
\usage{
extractProgramData(
  seurat_obj,
  nmf_obj,
  genes = NULL,
  cells = NULL,
  top_n_genes = 1000,
  reduction = "tsne",
  programID = "MergedProgram_01",
  ...
)
}
\arguments{
\item{seurat_obj}{A seurat object that holds the expresssion information and tSNE coordinates.}

\item{nmf_obj}{cNMF object from \code{identifyConsensusProgram}}

\item{genes}{The name(s) of genes to plot}

\item{cells}{The name(s) of cells to plot}

\item{top_n_genes}{Top n genes to use, if \code{genes} is not specified, default = 1000.}

\item{reduction}{String, the name of the reduction to use, default tsne.}

\item{programID}{Name of the program to check, default MergedProgram_01.}

\item{...}{Arguments for geom_point}
}
\value{
A tbl/df
}
\description{
Extract and scale the data for ploting
}
