% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{fill_extrapolate}
\alias{fill_extrapolate}
\title{Extrapolation of any NAs with previous value}
\usage{
fill_extrapolate(interventions, group_var = NULL, not = "itn_input_dist")
}
\arguments{
\item{interventions}{Site file interventions section}

\item{group_var}{Site grouping}

\item{not}{Character string naming Interventions not to be extrapolated}
}
\description{
Extrapolation of any NAs with previous value
}
