% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geneEnrichCelltypes}
\alias{geneEnrichCelltypes}
\title{Plot the enrichment for celltypes DEGs using the related genes of a given program}
\usage{
geneEnrichCelltypes(
  nmf_obj,
  programID = "MergedProgram_01",
  genes = NULL,
  top_n_genes = 1000
)
}
\arguments{
\item{nmf_obj}{}

\item{programID}{Name of the program to check, default MergedProgram_01.}

\item{genes}{A ranked and named genes vector. Default NULL.}

\item{top_n_genes}{Top n genes to use, if \code{genes} is not specified, default = 1000.}

\item{TERM2GENE}{A data.frame pass to \code{clusterProfiler::GSEA}}
}
\value{
A named list of ggplot objects
}
\description{
Plot the enrichment for celltypes DEGs using the related genes of a given program
}
