% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLowerTri2df}
\alias{getLowerTri2df}
\title{Get the lower triangle of a symmetric matrix}
\usage{
getLowerTri2df(x_mt, keep_diag = T)
}
\arguments{
\item{x_mt}{A symmetric matrix}

\item{keep_diag}{Logical. Keep diagonal values or not.}
}
\value{
A data.frame
}
\description{
Get the lower triangle of a symmetric matrix
}
