% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hrefs0}
\alias{hrefs0}
\title{internal function to read a stac query and get all its hrefs (every variable)}
\usage{
hrefs0(x)
}
\description{
internal function to read a stac query and get all its hrefs (every variable)
}
