% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identifyConsensusProgram}
\alias{identifyConsensusProgram}
\title{Identify consensus programs}
\usage{
identifyConsensusProgram(
  NMFs,
  n_reps,
  n_cores = 0,
  reps_ratio = 0.05,
  return_plot = F,
  verbose = T
)
}
\arguments{
\item{NMFs}{List of Results from repetitive NMF runs.}

\item{n_reps}{Integer. Number of replication for each K.}

\item{n_cores}{Integer. Number of cores to use, recommended to use multicore. Default = 0, all cores possible.}

\item{reps_ratio}{Double between \(0, 1\). Keeping clusters that have at least \code{round(n_reps * reps_ratio)} programs.}

\item{return_plot}{Logical. If return the plots for checking.}

\item{verbose}{Logical. If print the logs.}
}
\description{
Identify consensus programs
}
