% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identifyKnee}
\alias{identifyKnee}
\title{Identify knee}
\usage{
identifyKnee(x, y, direction = "down", draw_now = T, ...)
}
\arguments{
\item{x}{Numeric vector, x coordinates}

\item{y}{Numeric vector, y coordinates.}

\item{direction}{String, either 'up' or 'down'}

\item{draw_now}{Logical, if print the plot now}

\item{...}{Other parameters for kneer::create_knee_locator}
}
\value{
A numeric or NULL
}
\description{
Identify knee
}
