% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_df}
\alias{index_df}
\title{Index a data.frame by another}
\usage{
index_df(df, filter_row)
}
\arguments{
\item{df}{The input data.frame to obtain row indices for}

\item{filter_row}{The matching data.frame. Each row of `filter_row` is matched to a row in `df`.}
}
\description{
Take a data.frame `df` and return the row-index matching rows from another, `filter_row`.
}
