% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intervention_line_type}
\alias{intervention_line_type}
\title{Intervention line types}
\format{
## `intervention_line_type`
A named vector of length 15.
}
\usage{
intervention_line_type
}
\description{
Named lien types for intervention plotting
}
\keyword{datasets}
