% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irs_parameters}
\alias{irs_parameters}
\title{IRS efficacy parameters}
\format{
## `irs_parameters`
A data.frame
}
\usage{
irs_parameters
}
\description{
IRS efficacy parameters
}
\keyword{datasets}
