% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{last_used}
\alias{last_used}
\title{Select areas that have used an intervention in the last year with data}
\usage{
last_used(interventions, var, group_var)
}
\arguments{
\item{interventions}{Site file interventions section}

\item{var}{Interventions variable to check if last used}

\item{group_var}{Site grouping}
}
\description{
Select areas that have used an intervention in the last year with data
}
