% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{linear_interpolate}
\alias{linear_interpolate}
\title{Linearly interpolate NAs points}
\usage{
linear_interpolate(interventions, vars, group_var)
}
\arguments{
\item{interventions}{Site file interventions section}

\item{vars}{Character string naming Interventions to be interpolated}

\item{group_var}{Site grouping}
}
\description{
Linearly interpolate NAs points
}
