% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{link_resistance}
\alias{link_resistance}
\title{Add in resistance projections}
\usage{
link_resistance(interventions, pyrethroid_resistance, group_var)
}
\arguments{
\item{interventions}{Site file interventions section}

\item{pyrethroid_resistance}{Site file pyrethroid resistance}

\item{group_var}{Site grouping}
}
\description{
Add in resistance projections
}
