% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{link_vector_control_parameters}
\alias{link_vector_control_parameters}
\title{Link net type, irs type and net resistance to efficacy parameters}
\usage{
link_vector_control_parameters(interventions)
}
\arguments{
\item{interventions}{Site file interventions section}
}
\description{
Link net type, irs type and net resistance to efficacy parameters
}
