% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeCells}
\alias{mergeCells}
\title{Merge cells into metacell by clusters (deprecated)}
\usage{
mergeCells(
  expr_mt,
  celltypes,
  target_number = 30,
  K,
  k_beta = 2,
  double_check = F,
  ...
)
}
\arguments{
\item{expr_mt}{Expression matrix}

\item{celltypes}{should be a cell-id-named vector of cell types}

\item{target_number}{Number of cells to merge into one, default = 30.}

\item{k_beta}{Numeric}

\item{double_check}{Logical,}

\item{n_cores}{Use when double_check, default using all cores available}

\item{n_perm}{Use when double_check, default = 100}
}
\value{
A named list
}
\description{
Merge single cell data into meta cells. To be deprecated.
}
