% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_efficacy}
\alias{net_efficacy}
\title{Net efficacy parameters}
\format{
## `net_efficacy`
A data.frame
}
\usage{
net_efficacy
}
\description{
Linking efficacy with level of Pyrethroid resistance
}
\keyword{datasets}
