% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_interventions}
\alias{plot_interventions}
\title{Plot interventions}
\usage{
plot_interventions(
  interventions,
  population,
  group_var,
  country = TRUE,
  include = c("itn_use", "itn_input_dist", "fitted_usage", "tx_cov", "irs_cov",
    "rtss_cov", "smc_cov", "pmc_cov"),
  labels = c("ITN usage", "ITN model input", "ITN model usage", "Treatment", "IRS",
    "RTSS", "SMC", "PMC"),
  text_size = 8
)
}
\arguments{
\item{interventions}{Interventions input}

\item{population}{Population input}

\item{group_var}{Site grouping}

\item{country}{Aggregate to the country level}

\item{include}{Intervention variables to include}

\item{labels}{Labels for variables in include}

\item{text_size}{Size of font}
}
\value{
Intervention plot
}
\description{
Plot interventions
}
