% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_interventions_combined}
\alias{plot_interventions_combined}
\title{Plot multi panel interventions}
\usage{
plot_interventions_combined(
  interventions,
  population,
  group_var,
  text = c(11, 11)
)
}
\arguments{
\item{interventions}{Interventions input}

\item{population}{Population input}

\item{group_var}{Site grouping}

\item{text}{vector of text size for aggregated and disaggregated plots}
}
\value{
Combined country and admin plot
}
\description{
Plot multi panel interventions
}
