% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotEnrichCelltypes}
\alias{plotEnrichCelltypes}
\title{Plot the enrichment for celltypes DEGs using the related genes of a given program}
\usage{
plotEnrichCelltypes(
  nmf_obj,
  programID = "MergedProgram_01",
  TERM2GENE,
  all_types,
  genes = NULL,
  top_n_genes = 1000,
  maxGSSize = Inf,
  verbose = TRUE,
  GSEA_args = list(),
  enricher_args = list(),
  gseaplot_args = list()
)
}
\arguments{
\item{nmf_obj}{}

\item{programID}{Name of the program to check, default MergedProgram_01.}

\item{TERM2GENE}{A data.frame pass to \code{clusterProfiler::GSEA}}

\item{all_types}{A string of all the cell types.}

\item{genes}{A ranked and named genes vector. Default NULL.}

\item{top_n_genes}{Top n genes for the enricher, if \code{genes} is not specified. Default = 1000.}

\item{maxGSSize}{Default Inf}

\item{verbose}{Logical, whether to print log.}

\item{GSEA_args}{Other options that passed to \code{clusterProfiler::GSEA}, default empty.}

\item{enricher_args}{Other options that passed to \code{clusterProfiler::enricher}, default empty.}

\item{gseaplot_args}{Other options that passed to \code{clusterProfiler::gseaplot}, default empty.}
}
\value{
A named list of ggplot objects
}
\description{
Plot the enrichment for celltypes DEGs using the related genes of a given program
}
