% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotKnee}
\alias{plotKnee}
\title{Generate knee plot}
\usage{
plotKnee(plt_df, fit_obj)
}
\arguments{
\item{plt_df}{A data frame with x and y to plot.}

\item{fit_obj}{A knee object generated by \code{kneer}}
}
\value{
A ggplot plot
}
\description{
Generate knee plot
}
