% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotLoess}
\alias{plotLoess}
\title{Generate loess plot when knee cannot be found}
\usage{
plotLoess(plt_df, fit_obj)
}
\arguments{
\item{plt_df}{A data frame with x and y to plot.}

\item{fit_obj}{A loess fit object generated by \code{loess}}
}
\value{
A ggplot plot
}
\description{
Generate loess plot when knee cannot be found
}
