% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotProgramGenes}
\alias{plotProgramGenes}
\title{Plot related genes of a given program}
\usage{
plotProgramGenes(
  seurat_obj,
  nmf_obj,
  genes = NULL,
  top_n_genes = 1000,
  reduction = "tsne",
  plot_each = F,
  programID = "MergedProgram_01"
)
}
\arguments{
\item{seurat_obj}{A seurat object that holds the expresssion information and tSNE coordinates.}

\item{nmf_obj}{}

\item{genes}{The name(s) of genes to plot}

\item{top_n_genes}{Top n genes to use, if \code{genes} is not specified, default = 1000.}

\item{reduction}{Name of the reduction to use, default tsne.}

\item{plot_each}{}

\item{programID}{Name of the program to check, default MergedProgram_01.}
}
\value{
A named list of ggplot objects
}
\description{
Plot related genes of a given program
}
