% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processConsensusProgram}
\alias{processConsensusProgram}
\title{Process the consensus programs}
\usage{
processConsensusProgram(merged_program_mt, seurat_obj, program_cutoff = 0.1)
}
\arguments{
\item{merged_program_mt}{A matrix}

\item{seurat_obj}{A seurat object}

\item{program_cutoff}{Double, default = 0.1, between (0-1]}
}
\value{
A processed matrix
}
\description{
A wrapper function to process the consensus programs.
}
