% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processConsensusProgram2}
\alias{processConsensusProgram2}
\title{Process the consensus programs}
\usage{
processConsensusProgram2(merged_program_mt, sources, program_cutoff = 0.1)
}
\arguments{
\item{merged_program_mt}{A matrix}

\item{sources}{A named vector with the sources information}

\item{program_cutoff}{Double, default = 0.1, between (0-1]}
}
\value{
A processed matrix
}
\description{
A wrapper function to process the consensus programs.
}
