% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{Quantile normalization}
\usage{
quantile_normalize(mt, verbose = T)
}
\arguments{
\item{mt}{a matrix to norm by columns}

\item{verbose}{logical}
}
\description{
Quantile normalization
}
