% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runcNMF}
\alias{runcNMF}
\title{run cNMF to extract programs}
\usage{
runcNMF(
  merge_mt,
  k,
  n_reps,
  n_cores = 0,
  row_subset = 1,
  col_subset = 1,
  verbose = T
)
}
\arguments{
\item{merge_mt}{Matrix or sparse matrix. Normalized expression matrix of single-cell and bulk. Genes x Samples.}

\item{k}{Integer. K programs to run.}

\item{n_reps}{Integer. Number of replication for each K.}

\item{n_cores}{Integer. Number of cores to use, recommended to use multicore. Default = 0, all cores possible.}

\item{row_subset}{Double, \(0-1]. Proportion of rows to sample. Default = 1.}

\item{col_subset}{Double, \(0-1]. Proportion of columns to sample. Default = 1.}

\item{verbose}{Logical. If print the logs.}
}
\description{
run cNMF to extract programs
}
