% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_features_sc}
\alias{select_features_sc}
\title{Select features for single-cell data}
\usage{
select_features_sc(
  X,
  celltypes,
  batches = NULL,
  method = "wilcox",
  n_cores = 0
)
}
\arguments{
\item{X}{Matrix, genes x cells, a library-size-normalized matrix or a UMI count matrix or a TPM matrix.}

\item{celltypes}{Cell-named vector to specify}

\item{batches}{Cell-named vector to specify the batch information, default = NULL, not run}

\item{method}{String. Can be 'ttest' or 'wilcox'. Default = 'wilcox'}

\item{n_cores}{Integer. Default using all cores.}
}
\value{
A differential expression table.
}
\description{
Using differential expression to select feature genes for single-cell data.
}
