% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkConsensusProgram}
\alias{checkConsensusProgram}
\title{Check and plot the consensus programs}
\usage{
checkConsensusProgram(seurat_obj, program_mt, program_cutoff = 0.1)
}
\arguments{
\item{seurat_obj}{A seurat object}

\item{program_mt}{A program matrix}

\item{program_cutoff}{Double, default = 0.1, between (0-1]}
}
\value{
A named list
}
\description{
Check and plot the consensus programs
}
