% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determineK_runNMF}
\alias{determineK_runNMF}
\title{Determine K to use}
\usage{
determineK_runNMF(
  merge_mt,
  Ks2run,
  n_reps,
  n_cores = 0,
  draw_now = F,
  verbose = T,
  ...
)
}
\arguments{
\item{merge_mt}{Matrix or sparse matrix. Normalized expression matrix of single-cell and bulk. Genes x Samples.}

\item{Ks2run}{Integer vector. Ks to test.}

\item{n_reps}{Integer. Number of replication for each K.}

\item{n_cores}{Integer. Number of cores to use, recommended to use multicore. Default = 0, all cores possible.}

\item{draw_now}{Logical. If print out the plots.}

\item{verbose}{Logical. If print the logs.}
}
\description{
Determine K to use
}
