% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intervention_colours}
\alias{intervention_colours}
\title{Intervention colours}
\format{
## `intervention_colours`
A named vector of length 15.
}
\usage{
intervention_colours
}
\description{
Named colours for intervention plotting
}
\keyword{datasets}
