% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_interventions}
\alias{plot_interventions}
\title{Plot interventions}
\usage{
plot_interventions(
  interventions,
  population,
  group_var,
  include = c("itn_use", "itn_input_dist", "fitted_usage", "tx_cov", "irs_cov",
    "rtss_cov", "smc_cov", "pmc_cov", "lsm_cov"),
  labels = c("ITN usage", "ITN model input", "ITN model usage", "Treatment", "IRS",
    "RTSS", "SMC", "PMC", "LSM"),
  text_size = 8,
  facet_rows = 4
)
}
\arguments{
\item{interventions}{Interventions input}

\item{population}{Population input}

\item{group_var}{Site grouping}

\item{include}{Intervention variables to include}

\item{labels}{Labels for variables in include}

\item{text_size}{Size of font}

\item{facet_rows}{Number of rows for faceted plot}
}
\value{
Intervention plot
}
\description{
Plot interventions
}
