% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_interventions_combined}
\alias{plot_interventions_combined}
\title{Plot multi panel interventions}
\usage{
plot_interventions_combined(
  interventions,
  population,
  group_var,
  include = c("itn_use", "itn_input_dist", "fitted_usage", "tx_cov", "irs_cov",
    "rtss_cov", "smc_cov", "pmc_cov"),
  labels = c("ITN usage", "ITN model input", "ITN model usage", "Treatment", "IRS",
    "RTSS", "SMC", "PMC"),
  text = c(11, 11),
  facet_rows = 4
)
}
\arguments{
\item{interventions}{Interventions input}

\item{population}{Population input}

\item{group_var}{Site grouping}

\item{include}{Intervention variables to include}

\item{labels}{Labels for variables in include}

\item{text}{vector of text size for aggregated and disaggregated plots}

\item{facet_rows}{Number of rows for faceted plot}
}
\value{
Combined country and admin plot
}
\description{
Plot multi panel interventions
}
