% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessing}
\alias{preprocessing}
\title{Merge single-cell and bulk data}
\usage{
preprocessing(
  sc_data,
  bulk_data,
  celltypes,
  top_gene_ratio = 0.5,
  logarithmetic = T,
  scaling = T,
  quantile_normalization = F,
  downsample_size = NULL,
  draw_now = T,
  verbose = T,
  ...
)
}
\arguments{
\item{sc_data}{Sparse matrix of normalized counts or TPM, genes x cells}

\item{bulk_data}{Sparse matrix of TPM, genes x samples}

\item{celltypes}{Named vector, should be a cell-id-named vector of cell types}

\item{top_gene_ratio}{Numeric, top \% of HVG genes to remain}

\item{logarithmetic}{Logical, whether to log-transformed the data}

\item{scaling}{Logical, whether to scale data by Frobenius norm}

\item{quantile_normalization}{Logical, whether to perform quantile normalization}

\item{downsample_size}{}

\item{draw_now}{Logical, default = T}

\item{verbose}{Logical, default TRUE}
}
\value{
A named list
}
\description{
Merge single cells and bulks/spots into one expression matrix.
}
