% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{Quantile normalization}
\usage{
quantile_normalize(mt, n_cores = 30, verbose = T)
}
\arguments{
\item{mt}{Matrix to norm by columns}

\item{n_cores}{Number of cores to use. (to be added)}

\item{verbose}{Logical,  print the log.}
}
\value{
A quantile-normalized matrix.
}
\description{
Perform quantile normalization column-wisely on a matrix.
}
