% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{scene}
\alias{scene}
\title{Title}
\usage{
scene(
  x = cbind(146.614867, -43.298699),
  date = Sys.Date() + c(-14, -2),
  wh = c(5000),
  proj = NULL,
  res = max(c(10, wh/1024)),
  silent = FALSE,
  keep = 6L
)
}
\arguments{
\item{silent}{}
}
\description{
Title
}
