% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_features_sc}
\alias{select_features_sc}
\title{Select features for single-cell data}
\usage{
select_features_sc(
  X,
  celltypes,
  batches = NULL,
  method = "wilcox",
  n_cores = 0
)
}
\arguments{
\item{X}{Matrix, genes x cells, a library-size-normalized matrix or a UMI count matrix or a TPM matrix.}

\item{celltypes}{Named vector}

\item{batches}{Named vector}

\item{method}{String. Can be 'ttest' or 'wilcox'. Default = 'wilcox'}

\item{n_cores}{Integer. Default using all cores.}
}
\value{
A string vector of genes
}
\description{
Select features for single-cell data
}
