% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{set_change_point}
\alias{set_change_point}
\title{Set an intervention change point}
\usage{
set_change_point(interventions, var, year, target, sites = NULL)
}
\arguments{
\item{interventions}{Site file interventions section}

\item{var}{Interventions variable to modify}

\item{year}{Year of change point}

\item{target}{New value of change point}

\item{sites}{Site reference to modify at}
}
\description{
Modifies the site file interventions section with a new change point for an intervention
}
