% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{visualizeDiffContext}
\alias{visualizeDiffContext}
\title{Visualize different co-varying pattern in contexts
This function visualize the results of program pairs that are influenced significantly by the context, for each program pair, it includes the model and the summary}
\usage{
visualizeDiffContext(bulk_pg_dH_mt, res_lr, context_num_tbl, plt_dir = NULL)
}
\arguments{
\item{bulk_pg_dH_mt}{a matrix, with columns as programs, rows as bulk samples, calculated as \code{d} \%*\% \code{H}, where \code{d} and \code{H} are from \code{cNMF}}

\item{res_lr}{a list, result of \code{detectDiffContext}}

\item{context_num_tbl}{a tibble, result of \code{preprocess}}

\item{plt_dir}{a string specify the ploting directory, default NULL}
}
\value{
A list of plots

detectDiffContext
}
\description{
Visualize different co-varying pattern in contexts
This function visualize the results of program pairs that are influenced significantly by the context, for each program pair, it includes the model and the summary
}
